+++ src/main/java/task/Event.java
@@ -56,7 +56,7 @@ public class Event extends Task {
     public String toString() {
         int len = super.toString().length();
         return "[E]" + super.toString() + " (from: " + from.format(OUTPUT_FORMAT) + "\n"
-                + UI.UI.INDENT.repeat(3) + " ".repeat(len - 1)
+                + ui.Ui.INDENT.repeat(3) + " ".repeat(len - 1)
                 + "to: " + to.format(OUTPUT_FORMAT) + ")";
     }
 }
